// aiapp/frontend/src/pages/web-cloud/access/general/connections.types.ts
export interface ModemRouter {
  mode: 'router' | 'bridge';
  dmzEnabled: boolean;
  dmzIp?: string;
  firewallEnabled: boolean;
  upnpEnabled: boolean;
}

export interface ModemDns {
  primary: string;
  secondary: string;
}

export interface ModemMtu {
  size: number;
}